Added rudimentary ability for device to respond to receive indications
Indications sent from the BLE peripheral device are now detected. However, the program design is flawed because if the program is just allowed to run, the indication_task set in TestConnectedMenu's connect_to_device() is somehow blocked

Only when pushing buttons on the device does the task eventually finish. This will need to be improved in a future commit